fix(core): fix circular dependencies
