+++ .changeset/small-news-smash.md
@@ -1,0 +1,5 @@
+---
+'@chugsplash/core': patch
+---
+
+Remove circular dependencies
