+++ packages/core/package.json
@@ -34,8 +34,8 @@
     "@eth-optimism/contracts-bedrock": "^0.8.0",
     "@eth-optimism/core-utils": "^0.9.1",
     "@ethersproject/abstract-provider": "^5.7.0",
-    "core-js": "^3.27.1",
     "@openzeppelin/upgrades-core": "^1.20.6",
+    "core-js": "^3.27.1",
     "ethers": "^5.6.9",
     "fs": "^0.0.1-security",
     "handlebars": "^4.7.7",
