+++ packages/core/src/actions/index.ts
@@ -2,4 +2,4 @@ export * from './bundle'
 export * from './types'
 export * from './artifacts'
 export * from './execute'
-export * from './deployments'
+export * from './types'
