+++ packages/core/src/actions/types.ts
@@ -1,4 +1,15 @@
 import { BigNumber } from 'ethers'
+import { Fragment } from 'ethers/lib/utils'
+
+// TODO
+export type BuildInfo = any
+export type ContractArtifact = {
+  abi: Array<Fragment>
+  sourceName: string
+  contractName: string
+  bytecode: string
+}
+export type ContractASTNode = any
 
 /**
  * Possible action types.
