+++ packages/core/src/config/fetch.ts
@@ -84,3 +84,24 @@ export const verifyBundle = async (args: {
     bundle,
   }
 }
+
+/**
+ * Compiles a remote ChugSplashBundle from a uri.
+ *
+ * @param configUri URI of the ChugSplashBundle to compile.
+ * @param provider JSON RPC provider.
+ * @returns Compiled ChugSplashBundle.
+ */
+export const compileRemoteBundle = async (
+  configUri: string
+): Promise<{
+  bundle: ChugSplashActionBundle
+  canonicalConfig: CanonicalChugSplashConfig
+}> => {
+  const canonicalConfig = await chugsplashFetchSubtask({ configUri })
+
+  const bundle = await bundleRemote({
+    canonicalConfig,
+  })
+  return { bundle, canonicalConfig }
+}
