+++ packages/core/src/deployed.ts
@@ -3,7 +3,8 @@ import { getChainId } from '@eth-optimism/core-utils'
 import { Manifest } from '@openzeppelin/upgrades-core'
 import { Contract, providers } from 'ethers'
 
-import { CanonicalChugSplashConfig, chugsplashFetchSubtask } from './config'
+import { chugsplashFetchSubtask } from './config/fetch'
+import { CanonicalChugSplashConfig } from './config/types'
 import { getCanonicalConfigArtifacts, SolidityStorageLayout } from './languages'
 import {
   getChugSplashRegistry,
