+++ packages/core/src/index.ts
@@ -6,7 +6,6 @@ export * from './fund'
 export * from './constants'
 export * from './tasks'
 export * from './messages'
-export * from './deployments'
 export * from './execution'
 export * from './types'
 export * from './analytics'
