+++ packages/core/src/languages/solidity/storage.ts
@@ -1,7 +1,7 @@
 import { add0x, fromHexString, remove0x } from '@eth-optimism/core-utils'
 import { BigNumber, ethers, utils } from 'ethers'
 
-import { ParsedContractConfig } from '../../config'
+import { ParsedContractConfig } from '../../config/types'
 import {
   SolidityStorageLayout,
   SolidityStorageObj,
@@ -611,31 +611,4 @@ export const computeStorageSlots = (
   }
 
   return segments
-}
-
-export const addEnumMembersToStorageLayout = (
-  storageLayout: SolidityStorageLayout,
-  contractName: string,
-  sourceNodes: any
-): SolidityStorageLayout => {
-  // If no vars are defined or all vars are immutable, then storageLayout.types will be null and we can just return
-  if (storageLayout.types === null) {
-    return storageLayout
-  }
-
-  for (const layoutType of Object.values(storageLayout.types)) {
-    if (layoutType.label.startsWith('enum')) {
-      const canonicalVarName = layoutType.label.substring(5)
-      for (const contractNode of sourceNodes) {
-        if (contractNode.canonicalName === contractName) {
-          for (const node of contractNode.nodes) {
-            if (node.canonicalName === canonicalVarName) {
-              layoutType.members = node.members.map((member) => member.name)
-            }
-          }
-        }
-      }
-    }
-  }
-  return storageLayout
 }
