+++ packages/core/src/tasks/index.ts
@@ -7,18 +7,21 @@ import { create } from 'ipfs-http-client'
 import { ProxyABI } from '@chugsplash/contracts'
 
 import {
+  assertValidUpgrade,
   CanonicalChugSplashConfig,
   ChugSplashInput,
   ParsedChugSplashConfig,
+  readParsedChugSplashConfig,
+  verifyBundle,
 } from '../config'
 import {
   assertValidContracts,
-  assertValidUpgrade,
   computeBundleId,
   displayDeploymentTable,
   displayProposerTable,
   formatEther,
   generateFoundryTestArtifacts,
+  getBundleCompletionTxnHash,
   getChugSplashManager,
   getChugSplashManagerProxyAddress,
   getChugSplashRegistry,
@@ -29,7 +32,7 @@ import {
   isProjectRegistered,
   isTransparentProxy,
   isUUPSProxy,
-  readParsedChugSplashConfig,
+  readBuildInfo,
   registerChugSplashProject,
   setProxiesToReferenceNames,
   writeCanonicalConfig,
@@ -49,12 +52,9 @@ import {
   ChugSplashBundleState,
   ChugSplashBundleStatus,
   createDeploymentArtifacts,
-  proposeChugSplashBundle,
-  readBuildInfo,
 } from '../actions'
 import { getAmountToDeposit, getOwnerWithdrawableAmount } from '../fund'
 import { monitorExecution, postExecutionActions } from '../execution'
-import { getBundleCompletionTxnHash } from '../deployments'
 import { ChugSplashExecutorType, FoundryContractArtifact } from '../types'
 import {
   trackAddProposers,
@@ -65,6 +65,7 @@ import {
   trackFund,
   trackListProjects,
   trackListProposers,
+  trackProposed,
   trackRegistered,
   trackTransferProxy,
   trackWithdraw,
@@ -1563,4 +1564,72 @@ If you believe this is a mistake, please reach out to the developers or open an 
   )
 
   spinner.succeed('Proxy ownership successfully transferred to ChugSplash')
+}
+
+export const proposeChugSplashBundle = async (
+  provider: ethers.providers.JsonRpcProvider,
+  signer: ethers.Signer,
+  parsedConfig: ParsedChugSplashConfig,
+  bundle: ChugSplashActionBundle,
+  configUri: string,
+  remoteExecution: boolean,
+  ipfsUrl: string,
+  configPath: string,
+  spinner: ora.Ora = ora({ isSilent: true }),
+  confirm: boolean,
+  artifactPaths: ArtifactPaths,
+  canonicalConfigPath: string,
+  silent: boolean,
+  integration: Integration
+) => {
+  const projectName = parsedConfig.options.projectName
+
+  spinner.start(`Checking if the caller is a proposer...`)
+
+  spinner.succeed(`Caller is a proposer.`)
+
+  spinner.start(`Proposing ${projectName}...`)
+
+  const ChugSplashManager = getChugSplashManager(signer, projectName)
+
+  if (remoteExecution) {
+    await chugsplashCommitAbstractSubtask(
+      provider,
+      signer,
+      parsedConfig,
+      ipfsUrl,
+      true,
+      artifactPaths,
+      canonicalConfigPath,
+      integration,
+      spinner
+    )
+    // Verify that the bundle has been committed to IPFS with the correct bundle hash.
+    await verifyBundle({
+      configUri,
+      bundleId: computeBundleId(bundle.root, bundle.actions.length, configUri),
+      ipfsUrl,
+      artifactPaths,
+      integration,
+    })
+  }
+  // Propose the bundle.
+  await (
+    await ChugSplashManager.proposeChugSplashBundle(
+      bundle.root,
+      bundle.actions.length,
+      configUri,
+      await getGasPriceOverrides(provider)
+    )
+  ).wait()
+
+  const networkName = await resolveNetworkName(provider, integration)
+  await trackProposed(
+    await getProjectOwnerAddress(signer, projectName),
+    projectName,
+    networkName,
+    integration
+  )
+
+  spinner.succeed(`Proposed ${projectName}.`)
 }
